app/requirements.txt
fastapi==0.101.1
uvicorn==0.23.2
llama-cpp-python==0.1.78